+++ packages/client/src/observable/internal/list/ListCacheKey.ts
@@ -17,6 +17,7 @@
 import type { CollectionStorageData } from "../base-list/BaseCollectionQuery.js";
 import type { CacheKey } from "../CacheKey.js";
 import type { Canonical } from "../Canonical.js";
+import type { PivotInfo } from "../PivotCanonicalizer.js";
 import type { Rdp } from "../RdpCanonicalizer.js";
 import type { SimpleWhereClause } from "../SimpleWhereClause.js";
 import type { ListQuery } from "./ListQuery.js";
@@ -27,6 +28,8 @@ export const API_NAME_IDX = 1;
 export const WHERE_IDX = 2;
 export const ORDER_BY_IDX = 3;
 export const RDP_IDX = 4;
+export const INTERSECT_IDX = 5;
+export const PIVOT_IDX = 6;
 
 export interface ListStorageData extends CollectionStorageData {}
 
@@ -41,6 +44,10 @@ export interface ListCacheKey extends
       whereClause: Canonical<SimpleWhereClause>,
       orderByClause: Canonical<Record<string, "asc" | "desc" | undefined>>,
       rdpConfig?: Canonical<Rdp> | undefined,
+      intersectWith?:
+        | Canonical<Array<Canonical<SimpleWhereClause>>>
+        | undefined,
+      pivotInfo?: Canonical<PivotInfo> | undefined,
     ]
   >
 {