+++ packages/client/src/observable/internal/list/ListQueryOptions.ts
@@ -14,9 +14,24 @@
  * limitations under the License.
  */
 
+import type {
+  ObjectOrInterfaceDefinition,
+  SimplePropertyDef,
+  WhereClause,
+} from "@osdk/api";
 import type { CommonObserveOptions } from "../../ObservableClient/common.js";
 
-export interface ListQueryOptions extends CommonObserveOptions {
+export interface ListQueryOptions<
+  Q extends ObjectOrInterfaceDefinition = ObjectOrInterfaceDefinition,
+  RDPs extends Record<string, SimplePropertyDef> = Record<
+    string,
+    SimplePropertyDef
+  >,
+> extends CommonObserveOptions {
   pageSize?: number;
   autoFetchMore?: boolean | number;
+  intersectWith?: Array<{
+    where: WhereClause<Q, RDPs>;
+  }>;
+  pivotTo?: string;
 }