+++ benchmarks/tests/primary/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @osdk/benchmarks.primary
 
+## 0.1.0-beta.82
+
+### Patch Changes
+
+- Updated dependencies [a1ffb20]
+- Updated dependencies [e62c692]
+  - @osdk/client@2.6.0-beta.8
+
+## 0.1.0-beta.81
+
+### Patch Changes
+
+- Updated dependencies [507ec7f]
+  - @osdk/client@2.6.0-beta.7
+
 ## 0.1.0-beta.80
 
 ### Patch Changes