+++ packages/client/CHANGELOG.md
@@ -1,5 +1,32 @@
 # @osdk/client
 
+## 2.6.0-beta.8
+
+### Minor Changes
+
+- a1ffb20: Use new ephemeral media upload endpoint.
+- e62c692: Allow all applicable filters for struct sub properties to be applied on struct arrays
+
+### Patch Changes
+
+- Updated dependencies [a1ffb20]
+- Updated dependencies [e62c692]
+  - @osdk/generator-converters@2.6.0-beta.8
+  - @osdk/api@2.6.0-beta.8
+  - @osdk/client.unstable@2.6.0-beta.8
+
+## 2.6.0-beta.7
+
+### Minor Changes
+
+- 507ec7f: Allows top level filters with an implied and clause to support a combination of a not and and clause
+
+### Patch Changes
+
+- @osdk/api@2.6.0-beta.7
+- @osdk/client.unstable@2.6.0-beta.7
+- @osdk/generator-converters@2.6.0-beta.7
+
 ## 2.6.0-beta.6
 
 ### Minor Changes