+++ packages/foundry-sdk-generator/CHANGELOG.md
@@ -1,5 +1,32 @@
 # @osdk/foundry-sdk-generator
 
+## 2.6.0-beta.8
+
+### Minor Changes
+
+- a1ffb20: Use new ephemeral media upload endpoint.
+
+### Patch Changes
+
+- Updated dependencies [a1ffb20]
+- Updated dependencies [e62c692]
+  - @osdk/generator@2.6.0-beta.8
+  - @osdk/client@2.6.0-beta.8
+  - @osdk/api@2.6.0-beta.8
+
+## 2.6.0-beta.7
+
+### Minor Changes
+
+- f234d5c: Don't throw for nested array query params
+
+### Patch Changes
+
+- Updated dependencies [507ec7f]
+  - @osdk/client@2.6.0-beta.7
+  - @osdk/api@2.6.0-beta.7
+  - @osdk/generator@2.6.0-beta.7
+
 ## 2.6.0-beta.6
 
 ### Patch Changes