+++ packages/generator-converters/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @osdk/generator-converters
 
+## 2.6.0-beta.8
+
+### Minor Changes
+
+- a1ffb20: Use new ephemeral media upload endpoint.
+
+### Patch Changes
+
+- Updated dependencies [a1ffb20]
+- Updated dependencies [e62c692]
+  - @osdk/api@2.6.0-beta.8
+
+## 2.6.0-beta.7
+
+### Patch Changes
+
+- @osdk/api@2.6.0-beta.7
+
 ## 2.6.0-beta.6
 
 ### Minor Changes