+++ packages/widget.client/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @osdk/widget.client
 
+## 3.3.0-beta.6
+
+### Patch Changes
+
+- @osdk/widget.api@3.3.0-beta.6
+
+## 3.3.0-beta.5
+
+### Patch Changes
+
+- @osdk/widget.api@3.3.0-beta.5
+
 ## 3.3.0-beta.4
 
 ### Patch Changes