+++ packages/widget.vite-plugin/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @osdk/widget.vite-plugin
 
+## 3.3.0-beta.6
+
+### Patch Changes
+
+- @osdk/widget.api@3.3.0-beta.6
+
+## 3.3.0-beta.5
+
+### Minor Changes
+
+- 1940934: Support extension-less widget config import and improve reporting when config files not found
+- cc33fec: improve dev mode error surfacing
+- 54e674e: Fix widget dev mode with tailwind causing project imports
+- edb062f: Propagate validation error cause in message for widget config validation
+
+### Patch Changes
+
+- @osdk/widget.api@3.3.0-beta.5
+
 ## 3.3.0-beta.4
 
 ### Patch Changes